app/requirements.txt
amqp==2.5.1
billiard==3.6.1.0
boto3==1.9.234
botocore==1.12.234
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.40.0
django-filter==2.2.0
django-rest-auth==0.9.5
django-storages==1.7.2
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
docutils==0.15.2
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
kombu==4.6.4
MarkupSafe==1.1.1
more-itertools==7.2.0
oauthlib==3.1.0
psycopg2==2.8.3
Pygments==2.4.2
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
redis==3.3.8
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
zipp==0.6.0
